Early pixel-art shader support.
Edited rendering pipeline to use this new shader for post-processing.
SHADER DISABLED SO YOU READ THIS AND LEARN HOW TO TURN IT ON

Turn it on and off on the inspector of Settings/URP-HighFidelity-Renderer.asset (the default renderer), it's the Full Screen Pass Renderer feature you see in the inspector.

Edit the values of the shader on the inspector of Shaders/Pixel Material.mat (this is the shader material actually being passed onto the renderer)
Current values:
Resolution - Determines the pixel amount. Less resolution means less pixels.
Pixel Width/Height - Practically acts as an aspect ratio of sorts. Feel free to experiment with it.

The new "Shaders" folder contains three files.
-"Pixel.hlsl", where most of the "code & math" happens
- Pixel Material, which is a shader material passed onto the post-processing that contains all the shader settings.
- Pixel Shader, which is the Shader Graph (basically just a node system of logic, feel free to check it out if you want, dm me if you want an explanation on it)

Also added is the metadata Unity automatically makes when stuff is added, specifically for the new audio files Snek added, since I might as well.